+++ src/interfaces/publicationsIssn.js
@@ -31,6 +31,7 @@ import {graphql} from 'graphql';
 import schema from '../graphql';
 import HttpStatus from 'http-status';
 import {ApiError} from '@natlibfi/identifier-services-commons';
+import {hasAdminPermission, hasSystemPermission} from './utils';
 const objectId = require('mongodb').ObjectId;
 const date = new Date();
 
@@ -61,21 +62,25 @@ export default function () {
 		removeRequestISSN
 	};
 
-	async function createISSN(db, data) {
-		const query = `
-				mutation($input: InputPublicationIssn ) {
-					createPublicationIssn(input: $input) {
-						${queryReturn}
+	async function createISSN(db, data, user) {
+		if (hasAdminPermission(user)) {
+			const query = `
+					mutation($input: InputPublicationIssn ) {
+						createPublicationIssn(input: $input) {
+							${queryReturn}
+						}
 					}
-				}
-			`;
-		const args = {input: data};
-		const result = await graphql(schema, query, {createPublicationIssn}, db, args);
-		if (result.errors) {
-			throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+				`;
+			const args = {input: data};
+			const result = await graphql(schema, query, {createPublicationIssn}, db, args);
+			if (result.errors) {
+				throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+			}
+
+			return result;
 		}
 
-		return result;
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function createPublicationIssn({input}, db) {
 			const newPublication = {
@@ -92,20 +97,29 @@ export default function () {
 		}
 	}
 
-	async function readISSN(db, id) {
-		const query = `
+	async function readISSN(db, id, user) {
+		async function query() {
+			const query = `
 				{
 					publication_ISSN(id:${JSON.stringify(id)}) {
 						${queryReturn}
+						}
 					}
-				}
-			`;
-		const result = await graphql(schema, query, {publication_ISSN}, db);
-		if (result.data.publication_ISSN	 === null) {
-			throw new ApiError(HttpStatus.NOT_FOUND);
+					`;
+			const result = await graphql(schema, query, {publication_ISSN}, db);
+			if (result.data.publication_ISSN === null) {
+				throw new ApiError(HttpStatus.NOT_FOUND);
+			}
+
+			return result;
 		}
 
-		return result;
+		const response = query();
+		if (hasAdminPermission() || user.id === response.data.publication_ISSN.publisher) {
+			return response;
+		}
+
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function publication_ISSN({id}, db) {
 			const result = await db
@@ -115,21 +129,24 @@ export default function () {
 		}
 	}
 
-	async function updateISSN(db, id, data) {
-		const query = `
-				mutation($id:ID, $input: InputPublicationIssn ) {
-					updatePublicationIssn(id: $id, input: $input) {
-						${queryReturn}
-					}
+	async function updateISSN(db, id, values) {
+		const {data, user} = values;
+		if (hasAdminPermission(user) || hasSystemPermission(user)) {
+			const query = `
+			mutation($id:ID, $input: InputPublicationIssn ) {
+				updatePublicationIssn(id: $id, input: $input) {
+					${queryReturn}
 				}
+			}
 			`;
-		const args = {id: id, input: data};
-		const result = await graphql(schema, query, {updatePublicationIssn}, db, args);
-		if (result.errors) {
-			throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
-		}
+			const args = {id: id, input: data};
+			const result = await graphql(schema, query, {updatePublicationIssn}, db, args);
+			if (result.errors) {
+				throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+			}
 
-		return result;
+			return result;
+		}
 
 		async function updatePublicationIssn({input, id}, db) {
 			const updatePublication = {
@@ -175,20 +192,30 @@ export default function () {
 		}
 	}
 
-	async function queryISSN(db) {
-		const query = `
-				{
-					Publications_ISSN {
-						${queryReturn}
+	async function queryISSN(db, user) {
+		async function query() {
+			const query = `
+					{
+						Publications_ISSN {
+							${queryReturn}
+						}
 					}
-				}
-			`;
-		const result = await graphql(schema, query, {Publications_ISSN}, db);
-		if (result.errors) {
-			throw new Error();
+				`;
+			const result = await graphql(schema, query, {Publications_ISSN}, db);
+			if (result.errors) {
+				throw new Error();
+			}
+
+			return result;
 		}
 
-		return result;
+		const response = await query();
+
+		if (hasAdminPermission(user) || hasSystemPermission(user) || user.id === response.data.publication_ISSN.publisher) {
+			return response;
+		}
+
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function Publications_ISSN(root, db) {
 			const result = await db
@@ -199,21 +226,31 @@ export default function () {
 		}
 	}
 
-	async function createRequestISSN(db, data) {
-		const query = `
+	async function createRequestISSN(db, data, user) {
+		async function query() {
+			const query = `
 				mutation($input: InputPublicationRequestIssn) {
 					createPublicationRequestIssn(input: $input) {
-							${queryReturn}
+								${queryReturn}
+							}
 						}
-					}
-			`;
-		const args = {input: data};
-		const result = await graphql(schema, query, {createPublicationRequestIssn}, db, args);
-		if (result.errors) {
-			throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+				`;
+			const args = {input: data};
+			const result = await graphql(schema, query, {createPublicationRequestIssn}, db, args);
+			if (result.errors) {
+				throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+			}
+
+			return result;
 		}
 
-		return result;
+		const response = query();
+
+		if (hasSystemPermission(user) || user.id === response.data.updatePublicationRequestIssn.publisher) {
+			return response;
+		}
+
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function createPublicationRequestIssn({input}, db) {
 			const newPublicationRequest = {
@@ -230,21 +267,32 @@ export default function () {
 		}
 	}
 
-	async function updateRequestISSN(db, id, data) {
-		const query = `
-				mutation($id:ID, $input: InputPublicationRequestIssn) {
-					updatePublicationRequestIssn(id: $id, input: $input) {
-							${queryReturn}
-						}
-					}
+	async function updateRequestISSN(db, id, value) {
+		const {data, user} = value;
+		async function query() {
+			const query = `
+			mutation($id:ID, $input: InputPublicationRequestIssn) {
+				updatePublicationRequestIssn(id: $id, input: $input) {
+					${queryReturn}
+				}
+			}
 			`;
-		const args = {id: id, input: data};
-		const result = await graphql(schema, query, {updatePublicationRequestIssn}, db, args);
-		if (result.errors) {
-			throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+			const args = {id: id, input: data};
+			const result = await graphql(schema, query, {updatePublicationRequestIssn}, db, args);
+			if (result.errors) {
+				throw new ApiError(HttpStatus.UNPROCESSABLE_ENTITY);
+			}
+
+			return result;
 		}
 
-		return result;
+		const response = query();
+
+		if (hasAdminPermission(user) || hasSystemPermission(user) || user.id === response.data.updatePublicationRequestIssn.publisher) {
+			return response;
+		}
+
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function updatePublicationRequestIssn({id, input}, db) {
 			const updatePublicationRequest = {
@@ -267,20 +315,29 @@ export default function () {
 		}
 	}
 
-	async function readRequestISSN(db, id) {
-		const query = `
-				{
-					publicationRequest_ISSN(id:${JSON.stringify(id)}){
-						${queryReturn} state
-					}
+	async function readRequestISSN(db, id, user) {
+		async function query() {
+			const query = `
+			{
+				publicationRequest_ISSN(id:${JSON.stringify(id)}){
+					${queryReturn} state
 				}
+			}
 			`;
-		const result = await graphql(schema, query, {publicationRequest_ISSN}, db);
-		if (result.data.publicationRequest_ISSN	 === null) {
-			throw new ApiError(HttpStatus.NOT_FOUND);
+			const result = await graphql(schema, query, {publicationRequest_ISSN}, db);
+			if (result.data.publicationRequest_ISSN	 === null) {
+				throw new ApiError(HttpStatus.NOT_FOUND);
+			}
+
+			return result;
 		}
 
-		return result;
+		const response = query();
+		if (hasSystemPermission(user) || hasAdminPermission(user) || user.id === response.data.updatePublicationRequestIssn.publisher) {
+			return response;
+		}
+
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function publicationRequest_ISSN({id}, db) {
 			const result = await db
@@ -291,19 +348,23 @@ export default function () {
 	}
 
 	async function removeRequestISSN(db, id) {
-		const query = `
-				mutation{
-					deletePublicationRequestIssn(id:${JSON.stringify(id)}){
+		if (hasSystemPermission) {
+			const query = `
+			mutation{
+				deletePublicationRequestIssn(id:${JSON.stringify(id)}){
 					_id
 					}
 				}
 			`;
-		const result = await graphql(schema, query, {deletePublicationRequestIssn}, db);
-		if (result.errors) {
-			throw new ApiError(HttpStatus.NOT_FOUND);
+			const result = await graphql(schema, query, {deletePublicationRequestIssn}, db);
+			if (result.errors) {
+				throw new ApiError(HttpStatus.NOT_FOUND);
+			}
+
+			return result;
 		}
 
-		return result;
+		throw new ApiError(HttpStatus.FORBIDDEN);
 
 		async function deletePublicationRequestIssn({id}, db) {
 			const deletedPublicationRequest = await db
